Change after review, working with local storage
I made all the necessary changes after code review, completed the work with the local storage
