+++ src/components/Home/Home.module.scss
@@ -1,42 +1,46 @@
-@import "../../styles/core/import.scss";
+@import "../../styles/core";
+
+.Menu-button{
+    background-color: brown !important;
+}
  
- .Menu-open{
-    width: 255px !important;
-    background-color: #363740 !important;
-}
+//  .Menu-open{
+//     width: 255px !important;
+//     background-color: $color-menu-background !important;
+// }
 
-.Menu-header{
-    margin-left: 32px;
-    margin-top: 37px;
-    margin-bottom: 43px;
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-}
+// .Menu-header{
+//     margin-left: 32px;
+//     margin-top: 37px;
+//     margin-bottom: 43px;
+//     display: flex;
+//     flex-wrap: wrap;
+//     align-items: center;
+// }
 
-.Menu-header-text{
-    display: inline-block;
-    margin-left: 12px;
-    font-family: 'Mulish';
-    font-style: normal;
-    font-weight: 700;
-    font-size:  19px;
-    color: $color-menu-text;
-}
+// .Menu-header-text{
+//     display: inline-block;
+//     margin-left: 12px;
+//     font-family: 'Mulish';
+//     font-style: normal;
+//     font-weight: 700;
+//     font-size:  19px;
+//     color: $color-menu-text;
+// }
 
-.Menu-item-text{
-    font-family: 'Mulish';
-    font-style: normal;
-    font-weight: 400;
-    font-size:  16px;
-    color: $color-menu-text;
-    padding-left: 16px;
-}
+// .Menu-item-text{
+//     font-family: 'Mulish';
+//     font-style: normal;
+//     font-weight: 400;
+//     font-size:  16px;
+//     color: $color-menu-text;
+//     padding-left: 16px;
+// }
 
-.Menu-items:hover .Menu-item-text{
-    color: white;
-}
+// .Menu-items:hover .Menu-item-text{
+//     color: white;
+// }
 
-a{
-    text-decoration: none;
-}
+// a{
+//     text-decoration: none;
+// }
