+++ src/styles/core/color.scss
@@ -1,1 +1,3 @@
 $color-menu-text:#A4A6B3;
+$color-menu-background:#363740;
+$color-button: #3751FF;
